requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ORM & DB
SQLAlchemy==2.0.29
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.5

# Validation & settings
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Authentication & security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# File uploads (images)
python-multipart==0.0.9
Pillow==10.2.0
aiofiles==23.2.1

# Utils
email-validator==2.1.0.post1
python-dateutil==2.9.0.post0
loguru==0.7.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Optional (dev / quality)
httpx==0.27.0

dotenv